LayoutTests/media/video-end-expected.txt

TEST(!video.ended) OK
TEST(video.end == 0.5) OK
EVENT(ended)
TEST(video.currentTime == 0.5) OK
TEST(video.ended) OK
END OF TEST
